Fix failing tests when adding nodes out of order.
The assumption was that the first task for a new task_uuid would be the
start task. This is probably true in the real world but a test assumed
that the right thing would happen instead of creating a broken tree,
handling this is probably the right thing to do.